+++ api/src/test/scala/org/ensime/api/EnsimeTestData.scala
@@ -14,13 +14,12 @@ trait EnsimeTestData {
     }
   }
 
-  val typeInfo = new BasicTypeInfo("type1", 7, DeclaredAs.Method, "FOO.type1", List(), List(), None, Some(8))
+  val typeInfo = new BasicTypeInfo("type1", DeclaredAs.Method, "FOO.type1", List(), List(), None)
 
   val interfaceInfo = new InterfaceInfo(typeInfo, Some("DEF"))
-  val typeInspectInfo = new TypeInspectInfo(typeInfo, Some(1), List(interfaceInfo))
+  val typeInspectInfo = new TypeInspectInfo(typeInfo, List(interfaceInfo))
 
   val paramSectionInfo = new ParamSectionInfo(List(("ABC", typeInfo)), false)
-  val callCompletionInfo = new CallCompletionInfo(typeInfo, List(paramSectionInfo))
 
   val symFile = canon("/abc")
   val symbolDesignations = SymbolDesignations(
@@ -31,7 +30,7 @@ trait EnsimeTestData {
     )
   )
 
-  val symbolInfo = new SymbolInfo("name", "localName", None, typeInfo, false, Some(2))
+  val symbolInfo = new SymbolInfo("name", "localName", None, typeInfo, false)
 
   val implicitInfos = List(
     ImplicitConversionInfo(5, 6, symbolInfo),
@@ -46,9 +45,9 @@ trait EnsimeTestData {
 
   val packageInfo = new PackageInfo("name", "fullName", List())
 
-  val completionInfo = new CompletionInfo("name", new CompletionSignature(List(List(("abc", "def"), ("hij", "lmn"))), "ABC"), 88, false, 90, Some("BAZ"))
+  val completionInfo = new CompletionInfo("name", new CompletionSignature(List(List(("abc", "def"), ("hij", "lmn"))), "ABC", false), false, 90, Some("BAZ"))
 
-  val completionInfo2 = new CompletionInfo("name2", new CompletionSignature(List(List(("abc", "def"))), "ABC"), 90, true, 91, None)
+  val completionInfo2 = new CompletionInfo("name2", new CompletionSignature(List(List(("abc", "def"))), "ABC", false), true, 91, None)
 
   val completionInfoList = List(completionInfo, completionInfo2)
 
@@ -116,7 +115,7 @@ trait EnsimeTestData {
 
   val noteList = NewScalaNotesEvent(isFull = true, List(note1, note2))
 
-  val entityInfo: TypeInfo = new ArrowTypeInfo("Arrow1", 8, typeInfo, List(paramSectionInfo))
+  val entityInfo: TypeInfo = new ArrowTypeInfo("Arrow1", typeInfo, List(paramSectionInfo))
 
   val sourceFileInfo = SourceFileInfo(file1, Some("{/* code here */}"), Some(file2))
   val dtid = DebugThreadId(13)